Fix duplicate non-functional Image component in kit starters (SCB-1201)
Remove legacy Image Block rendering and prevent SXA/project Image double
registration in Available Renderings. Consolidate kit starters on project
Image datasource schema (image, caption) and align SYNC header sample data.
