+++ authoring/items/items/templates/items/ccl.powershell/Click Click Launch/Functions/Fix Duplicate Image Available Renderings.yml
@@ -1,0 +1,104 @@
+---
+ID: "e8f4a2c1-9b3d-4e7f-a1c2-3d5e6f7a8b9c"
+Parent: "4ff2646c-2b0f-428f-b9a0-d6ea531de713"
+Template: "dd22f1b3-bd87-4db2-9e7d-f7a496888d43"
+Path: "/sitecore/system/Modules/PowerShell/Script Library/Click Click Launch/Functions/Fix Duplicate Image Available Renderings"
+SharedFields:
+- ID: "b1a94ff0-6897-47c0-9c51-aa6acb80b1f0"
+  Hint: Script
+  Value: |
+    function Remove-RenderingGuidFromField {
+        param(
+            [Sitecore.Data.Items.Item]$Item,
+            [string]$FieldName,
+            [string]$GuidToRemove
+        )
+
+        $field = $Item.Fields[$FieldName]
+        if (-not $field -or [string]::IsNullOrEmpty($field.Value)) {
+            return $false
+        }
+
+        if ($field.Value -notlike "*$GuidToRemove*") {
+            return $false
+        }
+
+        $updatedValue = ($field.Value -split "`r?`n" | Where-Object {
+            -not [string]::IsNullOrWhiteSpace($_) -and ($_ -notmatch [regex]::Escape($GuidToRemove))
+        }) -join "`n"
+
+        if ($updatedValue -eq $field.Value) {
+            return $false
+        }
+
+        $Item.Editing.BeginEdit()
+        $field.Value = $updatedValue
+        $Item.Editing.EndEdit() | Out-Null
+        return $true
+    }
+
+    function Invoke-FixDuplicateImageAvailableRenderings {
+        [CmdletBinding()]
+        param(
+            [Parameter(Mandatory = $true, Position = 0)]
+            [Item]$Site
+        )
+
+        $sitePath = $Site.Paths.Path
+        $renderingImage = Get-Item -Path '/sitecore/layout/Renderings/Project/click-click-launch/Media/Image' -ErrorAction SilentlyContinue
+        if (-not $renderingImage) {
+            Write-Warning "Image rendering not found. Nothing to deduplicate."
+            return
+        }
+
+        $imageGuid = $renderingImage.ID.ToString()
+        $imageBlockGuid = "A563BE98-85FA-4DD5-94BD-9D9C937799F6"
+        $guidsToRemoveFromCategoryBuckets = @($imageGuid)
+        $guidsToRemoveEverywhere = @($imageBlockGuid)
+        $renderingListFieldNames = @("Renderings", "Allowed Controls")
+        $renderingCleanupPaths = @(
+            "$sitePath/Presentation/Available Renderings",
+            "$sitePath/Presentation/Placeholder Settings"
+        )
+
+        foreach ($renderingCleanupPath in $renderingCleanupPaths) {
+            if (-not (Test-Path $renderingCleanupPath)) { continue }
+
+            foreach ($renderingListItem in (Get-ChildItem -Path $renderingCleanupPath -Recurse -Language $Site.Language)) {
+                $isPrimaryBucket = $renderingListItem.Name -eq "click-click-launch"
+                $guidsToRemove = if ($isPrimaryBucket) { $guidsToRemoveEverywhere } else { $guidsToRemoveFromCategoryBuckets + $guidsToRemoveEverywhere }
+
+                foreach ($renderingListFieldName in $renderingListFieldNames) {
+                    foreach ($guidToRemove in $guidsToRemove) {
+                        if (Remove-RenderingGuidFromField -Item $renderingListItem -FieldName $renderingListFieldName -GuidToRemove $guidToRemove) {
+                            Write-Host "Removed $guidToRemove from $($renderingListItem.Paths.Path) — $renderingListFieldName"
+                        }
+                    }
+                }
+            }
+        }
+
+        Write-Host "Duplicate Image cleanup completed for $sitePath"
+    }
+Languages:
+- Language: en
+  Versions:
+  - Version: 1
+    Fields:
+    - ID: "25bed78c-4957-4165-998a-ca1b52f67497"
+      Hint: __Created
+      Value: 20260605T120000Z
+    - ID: "5dd74568-4d4b-44c1-b513-0af5f4cda34f"
+      Hint: __Created by
+      Value: |
+        sitecore\admin
+    - ID: "8cdc337e-a112-42fb-bbb4-4143751e123f"
+      Hint: __Revision
+      Value: "f1a2b3c4-d5e6-4789-a012-3456789abcde"
+    - ID: "badd9cf9-53e0-4d0c-bcc0-2d784c282f6a"
+      Hint: __Updated by
+      Value: |
+        sitecore\admin
+    - ID: "d9cf14b1-fa16-4ba6-9288-e8a174d4d522"
+      Hint: __Updated
+      Value: 20260605T120000Z
