+++ authoring/items/items/templates/items/ccl.templates.branches/click-click-launch/Components/Media/Available Headless Media Renderings/#name.yml
@@ -6,9 +6,7 @@ Path: "/sitecore/templates/Branches/Project/click-click-launch/Components/Media/
 SharedFields:
 - ID: "715ae6c0-71c8-4744-ab4f-65362d20ad65"
   Hint: Renderings
-  Value: |
-    {B2F8C9E3-4A7D-4E9B-8C1F-5D6A7B8C9D0E}
-    {7CB831E8-C889-4193-B822-A1131CB0B755}
+  Value: 
 Languages:
 - Language: en
   Versions:
