+++ authoring/items/items/templates/items/ccl.templates.branches/click-click-launch/SYNC/Add Partial Designs/Global/Header/Data/Image 1.yml
@@ -1,13 +1,17 @@
 ﻿---
 ID: "c6be30cd-04aa-4655-a942-4b00014a255f"
 Parent: "93160d7b-a085-4129-946b-425d131e0657"
-Template: "d885df8c-b2d6-4007-b34b-2bbafb527304"
+Template: "3720c1c9-5ef8-4ba1-828b-a4672503a1ae"
 Path: "/sitecore/templates/Branches/Project/click-click-launch/SYNC/Add Partial Designs/Global/Header/Data/Image 1"
 Languages:
 - Language: en
   Versions:
   - Version: 1
     Fields:
+    - ID: "14920a91-7250-436e-bf61-750d46eb7f3f"
+      Hint: image
+      Value: |
+        <image mediaid="9b0311dc-f62f-4db5-bac3-0040bbfc7370" />
     - ID: "25bed78c-4957-4165-998a-ca1b52f67497"
       Hint: __Created
       Value: 20250530T123225Z
@@ -15,18 +19,10 @@ Languages:
       Hint: __Owner
       Value: |
         sitecore\neli.kostadinova@sitecore.com
-    - ID: "57caf172-ce57-4e48-b3a1-46f8aea71c08"
-      Hint: Image
-      Value: |
-        <image mediaid="9b0311dc-f62f-4db5-bac3-0040bbfc7370" />
     - ID: "5dd74568-4d4b-44c1-b513-0af5f4cda34f"
       Hint: __Created by
       Value: |
         sitecore\neli.kostadinova@sitecore.com
-    - ID: "75fbf1b1-e7c5-494b-9633-693909f79425"
-      Hint: TargetUrl
-      Value: |
-        <link linktype="internal" id="585bf26e-d939-4265-ad84-79ad62d2a2bb" anchor="" querystring="" target="" class="" text="" title=""/>
     - ID: "8cdc337e-a112-42fb-bbb4-4143751e123f"
       Hint: __Revision
       Value: "1cc02873-8792-4bbe-9c23-249d44948471"
