+++ examples/kit-nextjs-article-starter/src/components/sxa/Image.tsx
@@ -1,76 +1,73 @@
-'use client';
+import type React from 'react';
+import { NextImage as ContentSdkImage, Text } from '@sitecore-content-sdk/nextjs';
+import { Default as ImageWrapper } from '@/components/image/ImageWrapper.dev';
+import { ImageProps } from '@/components/image/image.props';
+import { cn } from '@/lib/utils';
+import { NoDataFallback } from '@/utils/NoDataFallback';
 
-import {
-  Field,
-  ImageField,
-  NextImage as ContentSdkImage,
-  Link as ContentSdkLink,
-  LinkField,
-  useSitecore,
-} from '@sitecore-content-sdk/nextjs';
-import React, { JSX } from 'react';
+const DEFAULT_IMAGE_SIZES = '(max-width: 768px) 100vw, (max-width: 1024px) 90vw, 1200px';
+const BANNER_IMAGE_SIZES =
+  '(max-width: 640px) 100vw, (max-width: 768px) 768px, (max-width: 1024px) 1024px, (max-width: 1440px) 1280px, 1920px';
 
-interface Fields {
-  Image: ImageField & { metadata?: { [key: string]: unknown } };
-  ImageCaption: Field<string>;
-  TargetUrl: LinkField;
-}
+const getStyles = (params: ImageProps['params']) => params?.styles ?? params?.Styles ?? '';
 
-type ImageProps = {
-  params: { [key: string]: string };
-  fields: Fields;
-};
+export const Default: React.FC<ImageProps> = (props) => {
+  const fields = props.fields ?? props.rendering?.fields;
+  const { image, caption } = fields ?? {};
 
-export const Banner = (props: ImageProps): JSX.Element => {
-  const { page } = useSitecore();
-  const { Image } = props.fields;
-  const { TargetUrl } = props.fields;
-  const sxaStyles = props.params?.Styles ?? '';
-  const classNameList = `component image ${sxaStyles}`.trimEnd();
-
-  if (Image?.value && Image?.value.src) {
-    return (
-      <div className={classNameList}>
-        <div className="component-content">
-          {page.mode.isEditing || !props.fields.TargetUrl?.value?.href ? (
-            <ContentSdkImage field={Image} />
-          ) : (
-            <ContentSdkLink field={TargetUrl}>
-              <ContentSdkImage field={Image} />
-            </ContentSdkLink>
-          )}
-        </div>
-      </div>
-    );
+  if (fields === undefined) {
+    return <NoDataFallback componentName="Image" />;
   }
 
-  return <div className={classNameList}></div>;
+  return (
+    <figure className={cn('component', 'image', getStyles(props.params))}>
+      <ImageWrapper
+        image={image}
+        className="mb-[24px] h-full w-full object-cover"
+        sizes={DEFAULT_IMAGE_SIZES}
+      />
+      {caption && (
+        <figcaption className="image-caption field-imagecaption">
+          <Text field={caption} />
+        </figcaption>
+      )}
+    </figure>
+  );
 };
 
-export const Default = (props: ImageProps): JSX.Element => {
-  const { fields, params } = props;
-  const sxaStyles = params?.Styles ?? '';
-  const classNameList = `component image ${sxaStyles}`.trimEnd();
+export const Banner: React.FC<ImageProps> = (props) => {
+  const fields = props.fields ?? props.rendering?.fields;
+  const { image } = fields ?? {};
+  const id = props.params?.RenderingIdentifier;
 
-  if (fields) {
-    const { Image } = props.fields;
-
-    const modifyImageProps = {
-      ...Image,
-      value: {
-        ...Image?.value,
-        alt: Image?.value?.alt || 'image',
-      },
-    };
-
-    return (
-      <div className={classNameList}>
-        <div className="component-content">
-          <ContentSdkImage field={modifyImageProps} />
-        </div>
-      </div>
-    );
+  if (fields === undefined) {
+    return <NoDataFallback componentName="Image" />;
   }
 
-  return <div className={classNameList}></div>;
+  const imageField = image && {
+    ...image,
+    value: {
+      ...image.value,
+      style: { objectFit: 'cover', width: '100%', height: '100%' },
+    },
+  };
+
+  const altText = typeof image?.value?.alt === 'string' ? image.value.alt : 'Hero banner';
+
+  return (
+    <figure
+      className={cn('component hero-banner', getStyles(props.params))}
+      id={typeof id === 'string' ? id : undefined}
+    >
+      <div className="component-content sc-sxa-image-hero-banner">
+        <ContentSdkImage
+          field={imageField}
+          loading="eager"
+          fetchPriority="high"
+          sizes={BANNER_IMAGE_SIZES}
+          alt={altText}
+        />
+      </div>
+    </figure>
+  );
 };
