+++ examples/kit-nextjs-location-finder/src/__tests__/components/sxa/Image.test.tsx
@@ -2,112 +2,108 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { Default as Image, Banner } from '@/components/sxa/Image';
-import { Page } from '@sitecore-content-sdk/nextjs';
+import type { ImageProps } from '@/components/image/image.props';
+import type { ImageField, Page } from '@sitecore-content-sdk/nextjs';
 
-// Mock Sitecore SDK
-jest.mock('@sitecore-content-sdk/nextjs', () => ({
-  useSitecore: jest.fn(() => ({
-    page: {
-      mode: {
-        isEditing: false,
-        isPreview: false,
-        isNormal: true,
-      },
+const mockPage = {
+  mode: {
+    isEditing: false,
+    isPreview: false,
+    isNormal: true,
+    name: 'normal' as const,
+    designLibrary: { isVariantGeneration: false },
+    isDesignLibrary: false,
+  },
+  layout: {
+    sitecore: {
+      context: {},
+      route: null,
     },
-  })),
-  NextImage: jest.fn(({ field }) => (
-    // eslint-disable-next-line @next/next/no-img-element
-    <img data-testid="sxa-image" src={field?.value?.src} alt={field?.value?.alt} />
-  )),
-  Link: jest.fn(({ field, children }) => (
-    <a href={field?.value?.href} data-testid="image-link">
-      {children}
-    </a>
-  )),
-  Text: jest.fn(({ field, className }) => (
+  },
+  locale: 'en',
+} as Page;
+
+jest.mock('@/lib/utils', () => ({
+  cn: (...classes: string[]) => classes.filter(Boolean).join(' '),
+}));
+
+jest.mock('@sitecore-content-sdk/nextjs', () => ({
+  Text: ({ field, className }: { field?: { value?: string }; className?: string }) => (
     <span className={className} data-testid="image-caption">
       {field?.value}
     </span>
-  )),
+  ),
+  NextImage: ({ field, alt }: { field?: ImageField; alt?: string }) => (
+    // eslint-disable-next-line @next/next/no-img-element
+    <img
+      data-testid="next-image"
+      src={field?.value?.src as string | undefined}
+      alt={(alt ?? field?.value?.alt) as string | undefined}
+    />
+  ),
 }));
 
-describe('SXA Image', () => {
-  const mockPage = {
-    mode: {
-      isEditing: false,
-      isPreview: false,
-      isNormal: true,
-      name: 'normal' as const,
-      designLibrary: { isVariantGeneration: false },
-      isDesignLibrary: false,
-    },
-    layout: {
-      sitecore: {
-        context: {},
-        route: null,
-      },
-    },
-    locale: 'en',
-  } as Page;
+jest.mock('@/components/image/ImageWrapper.dev', () => ({
+  Default: ({ image, className }: { image?: ImageField; className?: string }) => (
+    <div data-testid="image-wrapper" className={className}>
+      {/* eslint-disable-next-line @next/next/no-img-element */}
+      <img
+        src={image?.value?.src as string | undefined}
+        alt={image?.value?.alt as string | undefined}
+      />
+    </div>
+  ),
+}));
 
-  const mockFields = {
-    Image: {
-      value: {
-        src: '/images/alaris-type-3-ambulance.jpg',
-        alt: 'Alaris Type III Ambulance Front View',
+jest.mock('@/utils/NoDataFallback', () => ({
+  NoDataFallback: ({ componentName }: { componentName?: string }) => (
+    <div data-testid="no-data-fallback">{componentName}</div>
+  ),
+}));
+
+describe('Image Component', () => {
+  const defaultProps: ImageProps = {
+    params: { styles: 'image-styles', RenderingIdentifier: 'image-1' },
+    fields: {
+      image: {
+        value: {
+          src: '/images/alaris-type-3-ambulance.jpg',
+          alt: 'Alaris Type III Ambulance Front View',
+        },
+      },
+      caption: {
+        value: 'Front view of the Alaris Type III Ambulance',
       },
     },
-    ImageCaption: {
-      value: 'Type III Ambulance - Extended Storage Configuration',
-    },
-    TargetUrl: {
-      value: {
-        href: '/vehicles/ambulances/type-3/details',
-      },
-    },
+    page: mockPage,
+    rendering: { componentName: 'Image' },
+    componentMap: new Map(),
   };
 
-  it('renders image with caption and link', () => {
-    render(<Image params={{ styles: '', RenderingIdentifier: 'img-1' }} fields={mockFields} page={mockPage} />);
+  it('renders image and caption', () => {
+    render(<Image {...defaultProps} />);
 
-    expect(screen.getByTestId('sxa-image')).toHaveAttribute(
-      'src',
-      '/images/alaris-type-3-ambulance.jpg'
-    );
+    expect(screen.getByTestId('image-wrapper')).toBeInTheDocument();
     expect(screen.getByTestId('image-caption')).toHaveTextContent(
-      'Type III Ambulance - Extended Storage Configuration'
-    );
-    expect(screen.getByTestId('image-link')).toHaveAttribute(
-      'href',
-      '/vehicles/ambulances/type-3/details'
+      'Front view of the Alaris Type III Ambulance'
     );
   });
 
-  it('renders image without link when not provided', () => {
-    const fieldsWithoutLink = {
-      ...mockFields,
-      TargetUrl: { value: { href: '' } },
-    };
+  it('renders banner variant', () => {
+    const { container } = render(<Banner {...defaultProps} />);
 
+    expect(container.querySelector('.hero-banner')).toBeInTheDocument();
+    expect(screen.getByTestId('next-image')).toBeInTheDocument();
+  });
+
+  it('renders fallback when fields are undefined', () => {
     render(
-      <Image params={{ styles: '', RenderingIdentifier: 'img-2' }} fields={fieldsWithoutLink} page={mockPage} />
-    );
-
-    expect(screen.getByTestId('sxa-image')).toBeInTheDocument();
-    expect(screen.queryByTestId('image-link')).not.toBeInTheDocument();
-  });
-
-  it('renders Banner variant with background image', () => {
-    const { container } = render(
-      <Banner
-        params={{ styles: 'hero-section', RenderingIdentifier: 'banner-1' }}
-        fields={mockFields}
-        page={mockPage}
+      <Image
+        {...defaultProps}
+        fields={undefined as unknown as ImageProps['fields']}
       />
     );
 
-    const bannerDiv = container.querySelector('.hero-banner');
-    expect(bannerDiv).toBeInTheDocument();
-    expect(bannerDiv).toHaveClass('hero-section');
+    expect(screen.getByTestId('no-data-fallback')).toHaveTextContent('Image');
   });
 });
